requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scipy>=1.7.0
scikit-image>=0.19.0
opencv-python>=4.5.0
pillow>=9.0.0

# FastAPI and server
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# Image processing and metadata
exifread>=3.0.0
piexif>=1.1.3

# Data processing
pandas>=1.3.0
matplotlib>=3.5.0

# Optional ML dependencies
transformers>=4.20.0
accelerate>=0.20.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
requests>=2.28.0
